requirements.txt
google-auth-oauthlib>=0.4.2
google-api-python-client>=1.12.8
pystray
keyring